requirements.txt
Flask~=3.0.3
python-dotenv~=1.2.1
flask-restx~=1.3.0